Правки по замечаниям код ревьюера
