+++ src/services/actions/tabs.js
@@ -1,0 +1,6 @@
+import { SET_CURRENT_TAB } from '../constants';
+
+export const setCurrentTab = (value) => ({
+  type: SET_CURRENT_TAB,
+  value: value
+});
