+++ src/services/constants.js
@@ -17,5 +17,3 @@ export const ORDER_REGISTRATION = 'ORDER_REGISTRATION';
 export const ORDER_REGISTRATION_SUCCESS = 'ORDER_REGISTRATION_SUCCESS';
 export const ORDER_REGISTRATION_FAILED = 'ORDER_REGISTRATION_FAILED';
 export const ORDER_CLEAN = 'ORDER_CLEAN';
-
-export const IS_DRAG_INGREDIENT = 'IS_DRAG_INGREDIENT';
