+++ src/services/reducers/tabs.js
@@ -1,0 +1,35 @@
+import {
+  SET_CURRENT_TAB
+} from '../constants';
+
+
+const initialState = [
+  {
+    value: 'bun',
+    text: 'Булки',
+    isActive: true,
+  },
+  {
+    value: 'sauce',
+    text: 'Соусы',
+    isActive: false,
+  },
+  {
+    value: 'main',
+    text: 'Начинки',
+    isActive: false,
+  },
+];
+
+const tabs = (state = initialState, action) => {
+  switch (action.type) {
+    case SET_CURRENT_TAB: {
+      return state.map((tab) => ({ ...tab, isActive: tab.value === action.value }));
+    }
+    default: {
+      return state;
+    }
+  }
+}
+
+export default tabs;
