+++ src/services/selectors.js
@@ -1,0 +1,9 @@
+export const selectorTabs = state => state.tabs;
+
+export const selectorIngredients = state => state.ingredients;
+
+export const selectorSelectedIngredients = state => state.selectedIngredients;
+
+export const selectorDetailsIngredient = state => state.detailsIngredient;
+
+export const selectorOrder = state => state.order;
